split internal logic into client class, removed namespace
