+++ src/eventarc/cloudevent.ts
@@ -1,5 +1,6 @@
 
 /*!
+ * @license
  * Copyright 2022 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -37,5 +38,5 @@ export interface CloudEvent {
    data?: object | string;
  
    // Custom attributes/extensions.
-   [key: string]: unknown;
+   [key: string]: any;
  }
