+++ src/eventarc/index.ts
@@ -1,5 +1,5 @@
 /*!
- * @@license
+ * @license
  * Copyright 2022 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
